[DF-7621] Removed skybox from editor
Former-commit-id: fdaf16d0ea0c3b1fdfc481bf7e902b27351ba4e5

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/TextDrawSystem.cpp
@@ -100,8 +100,8 @@ void TextDrawSystem::Draw()
 
 void TextDrawSystem::PushNextBatch(const Color& color)
 {
-	size_t vertexCount = vertices.size();
-	size_t indexCount = 6 * vertexCount / 4;
+	uint32 vertexCount = static_cast<uint32> (vertices.size());
+	uint32 indexCount = 6 * vertexCount / 4;
 
 	RenderSystem2D::BatchDescriptor batchDescriptor;
     batchDescriptor.singleColor = color;
